feat: request admin page

+++ src/Command/UpdateMoviesCommand.php
@@ -6,32 +6,34 @@ use App\Entity\MovieBackdrop;
 use App\Entity\Movie;
 use App\Repository\MovieRepository;
 use App\Repository\RequestRepository;
+use App\Service\MovieService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 class UpdateMoviesCommand extends Command
 {
     protected static $defaultName = 'app:update-movies';
 
-    private HttpClientInterface $tmdbClient;
     private EntityManagerInterface $entityManager;
     private MovieRepository $movieRepository;
     private RequestRepository $requestRepository;
+    private MovieService $movieService;
 
     public function __construct(
-        HttpClientInterface $tmdbClient,
         EntityManagerInterface $entityManager,
         MovieRepository $movieRepository,
-        RequestRepository $requestRepository
+        RequestRepository $requestRepository,
+        MovieService $movieService
     ) {
         parent::__construct();
-        $this->tmdbClient = $tmdbClient;
         $this->movieRepository = $movieRepository;
         $this->entityManager = $entityManager;
         $this->requestRepository = $requestRepository;
+        $this->movieService = $movieService;
     }
 
     /**
@@ -41,10 +43,14 @@ class UpdateMoviesCommand extends Command
     {
         foreach ($this->movieRepository->findAll() as $movie) {
             if (!file_exists('../movies/' . $movie->getId() . '.mp4')) {
-                $this->entityManager->remove($movie);
-                echo 'Removed orphaned movie ' . $movie->getId() . ' -> ' . $movie->getTitle() . "\n";
+                if (!$movie->getIsHidden()) {
+                    $this->entityManager->remove($movie);
+                    echo 'Removed orphaned movie ' . $movie->getId() . ' -> ' . $movie->getTitle() . "\n";
+                }
             }
         }
+
+        $this->entityManager->flush();
 
         try {
             $files = scandir('../movies');
@@ -52,78 +58,38 @@ class UpdateMoviesCommand extends Command
             foreach ($files as $file) {
                 if (preg_match('/((\d+)\.mp4)/', $file, $matches)) {
                     $id = (int) $matches[0];
+                    $movie = $this->movieRepository->find($id);
 
-                    if (!$this->movieRepository->find($id)) {
-                        $info = $this->fetchMovieInfo($id);
+                    if (null === $movie) {
+                        $movie = $this->movieService->findById($id);
+                    }
 
-                        $infoMovie = $info['movie'];
-
-                        $movie = new Movie();
-                        $movie->setId($id);
-                        $movie->setTitle($infoMovie['title']);
-                        $movie->setPoster($infoMovie['poster_path']);
-                        $movie->setYear((int) (substr($infoMovie['release_date'], 0, 4)));
+                    if ($movie->getIsHidden()) {
+                        $movie->setIsHidden(false);
                         $this->entityManager->persist($movie);
 
-                        // Filter backdrops, we do not want any backdrops with translated titles
-                        $backdrops = $info['images']['backdrops'] ?? [];
+                        echo 'Registered movie ' . $movie->getId() . ' -> ' . $movie->getTitle() . ' ( ' . count($movie->getBackdrops()) . ' Backdrops)' . "\n";
 
-                        $backdrops = array_filter($backdrops, static function ($backdrop) {
-                            return $backdrop['iso_639_1'] === null;
-                        });
+                        $requests = $this->requestRepository->findBy([
+                            'movie' => $movie,
+                        ]);
 
-                        $backdrops = array_map(static function($backdrop) {
-                            return $backdrop['file_path'];
-                        }, $backdrops);
-
-                        // Persist backdrops in database
-                        foreach ($backdrops as $backdropPath) {
-                            $backdrop = new MovieBackdrop();
-                            $backdrop->setMovie($movie);
-                            $backdrop->setFile($backdropPath);
-
-                            $this->entityManager->persist($backdrop);
+                        if ($requests) {
+                            foreach ($requests as $request) {
+                                $this->entityManager->remove($request);
+                            }
                         }
-
-                        $request = $this->requestRepository->find($id);
-                        if ($request) {
-                            $this->entityManager->remove($request);
-                        }
-
-                        echo 'Registered movie ' . $movie->getId() . ' -> ' . $movie->getTitle() . ' ( ' . count($backdrops) . ' Backdrops)' . "\n";
                     }
                 }
             }
-        } catch(\ErrorException $ignored) {
-            echo 'Folder missing: \'movies\'';
+        } catch(\ErrorException $ex) {
+            echo $ex->getMessage().'<br/>';
+            foreach($ex->getTrace() as $trace) {
+                echo $trace.'<br/>';
+            }
         }
 
         $this->entityManager->flush();
         return Command::SUCCESS;
-    }
-
-    /**
-     * @return array
-     */
-    private function fetchMovieInfo(int $tmdbId)
-    {
-        $movieInfo = json_decode(
-            $this->tmdbClient->request('GET', "movie/$tmdbId", [
-                'query' => [
-                    'language' => 'de'
-                ]
-            ])->getContent(),
-            true
-        );
-
-        $movieImages = json_decode(
-            $this->tmdbClient->request('GET', "movie/$tmdbId/images")->getContent(),
-            true
-        );
-
-        return [
-            'movie' => $movieInfo,
-            'images' => $movieImages,
-        ];
     }
 }
